+++ epictrack-web/src/components/workPlan/WorkPlanContainer.tsx
@@ -70,8 +70,9 @@ const WorkPlanContainer = () => {
           <ETTab
             label="Status"
             icon={
-              (ctx.statuses.length === 0 ||
-                ctx.statuses[0].is_approved === false) && <IndicatorIcon />
+              (ctx.statuses.length === 0 || ctx.isStatusOutOfDate()) && (
+                <IndicatorIcon />
+              )
             }
           />
           <ETTab label="Issues" />
