+++ epictrack-web/src/components/workPlan/WorkPlanContext.tsx
@@ -8,7 +8,7 @@ import {
 } from "react";
 import { useSearchParams } from "../../hooks/SearchParams";
 import workService from "../../services/workService/workService";
-import { Work, WorkPhase, WorkPhaseAdditionalInfo } from "../../models/work";
+import { Work, WorkPhaseAdditionalInfo } from "../../models/work";
 import { StaffWorkRole } from "../../models/staff";
 import {
   ACTIVE_STATUS,
@@ -20,6 +20,7 @@ import { Status } from "../../models/status";
 import { WorkIssue } from "../../models/Issue";
 import statusService from "../../services/statusService/statusService";
 import dateUtils from "../../utils/dateUtils";
+import moment from "moment";
 
 interface WorkplanContextProps {
   selectedWorkPhase?: WorkPhaseAdditionalInfo;
@@ -40,6 +41,7 @@ interface WorkplanContextProps {
   getWorkStatuses: () => Promise<void>;
   issues: WorkIssue[];
   setIssues: Dispatch<SetStateAction<WorkIssue[]>>;
+  isStatusOutOfDate: () => boolean;
 }
 interface WorkPlanContainerRouteParams extends URLSearchParams {
   work_id: string;
@@ -62,6 +64,7 @@ export const WorkplanContext = createContext<WorkplanContextProps>({
   issues: [],
   setIssues: () => ({}),
   getWorkStatuses: () => new Promise((resolve) => resolve),
+  isStatusOutOfDate: () => false,
 });
 
 export const WorkplanProvider = ({
@@ -169,6 +172,25 @@ export const WorkplanProvider = ({
     }
   };
 
+  const STATUS_DATE_THRESHOLD = 7;
+
+  const isStatusOutOfDate = () => {
+    const lastApprovedStatus = statuses.find((status) => status.is_approved);
+
+    if (!lastApprovedStatus) {
+      return false;
+    }
+
+    const daysAgo = moment().subtract(STATUS_DATE_THRESHOLD, "days");
+    const NDaysAgo = dateUtils.diff(
+      daysAgo.toLocaleString(),
+      lastApprovedStatus?.posted_date,
+      "days"
+    );
+
+    return NDaysAgo > 0;
+  };
+
   return (
     <WorkplanContext.Provider
       value={{
@@ -188,6 +210,7 @@ export const WorkplanProvider = ({
         statuses,
         issues,
         setIssues,
+        isStatusOutOfDate,
       }}
     >
       {children}
